fim de aula

+++ aula20160908/ptr4.c
@@ -2,25 +2,25 @@
 #include <stdlib.h>
 int main ()
 {
-    int num1, num2, soma;
+    float num1, num2, soma;
     int cont = 0;
     //int tam = sizeof(vetor)/sizeof(int);
 
     printf(" Digite dois numeros reais\n");
-    scanf("%d",& num1);
-    scanf("%d",& num2);
+    scanf("%f",& num1);
+    scanf("%f",& num2);
 
 
 
 
     //unsigned char *p= NULL, *q; //inicializar ponteiros
 
-    printf("\n %p: %d\n", &num1, num1);
-    printf("\n %p: %d\n", &num2, num2);
+    printf("\n %p: %f\n", &num1, num1);
+    printf("\n %p: %f\n", &num2, num2);
 
     soma=num1 + num2;
-    printf("Soma e:%d\n", soma);
-    printf("%p: %d\n", &soma,soma);
+    printf("Soma e:%f\n", soma);
+    printf("%p: %f\n", &soma,soma);
